Removed the composer vendor folder - added a new dependency named spatie/schema
Removed the composer's vendor folder, this folder contains the project's dependencies. We do not need to keep this in our repository for several reasons as it is suggested in this post:

https://getcomposer.org/doc/faqs/should-i-commit-the-dependencies-in-my-vendor-directory.md

I updated the installation section of the README.md file explaining how a user can install our plugin, the user has to now run commands from composer to downlaod dependencies (vendor folder).

Finaly i added a new dependency named spatie/schema, with this maybe we can enhance the way we generate schema metadata.