returned vendor direcotry
I have deleted vendor folder from .gitignore, so now vendor folder is also commited to VCS and it does not require additional actions from user. Test with relize later if it is possible to have vendor foldr only in relizes to decrease overheading of repository and commits. If possible, vendor will be included to .gitignore again.